+++ encryption/encryptedtext_to_data.py
@@ -11,7 +11,8 @@ def encryptedtext_to_data(encryptedtext: bytearray, password: bytearray, *,
                           pin: Optional[bytearray] = None, 
                           iterations: int = 100_000,
                           Nmin: int = 100_000,
-                          Nmax: int = 10_000_000
+                          Nmax: int = 10_000_000,
+                          delete_keys: bool = True
                           ) -> bytearray:
     """
     encryptedtext_to_data decrypts the encrypted text to generate the data.
@@ -40,6 +41,9 @@ def encryptedtext_to_data(encryptedtext: bytearray, password: bytearray, *,
         Nmax : int
             If pin is not None, The maximum number of iteration to derive the key using PBKDF2.
             The default is 100_000.
+        delete_keys : bool
+            Delete the password and the pin correctly from memory at the end of the function.
+            Default is True.
 
     Returns
     -------
@@ -54,7 +58,7 @@ def encryptedtext_to_data(encryptedtext: bytearray, password: bytearray, *,
             If iterations <= 0 or not (0 < Nmin < Nmax)
             If encryptedtext is not at least 80 bytes long 
         WrongKeyError
-            If the password or the pin is incorect.
+            If the password or the pin is incorrect.
             If the encryptedtext has been modified.
     """
     if (not isinstance(encryptedtext, bytearray)) or (not isinstance(password, bytearray)) or ((pin is not None) and (not isinstance(pin, bytearray))):
@@ -69,6 +73,8 @@ def encryptedtext_to_data(encryptedtext: bytearray, password: bytearray, *,
         raise TypeError("Parameters Nmin and Nmax not respect 0 < Nmin < Nmax")
     if len(encryptedtext) < 80:
         raise ValueError("Parameter encryptedtext is not at least 80 bytes long.")
+    if not isinstance(delete_keys, bool):
+        raise ValueError("Parameter delete_keys is not a booleen.")
     # Encryption
     if pin is not None:
         iterations = compute_iteration_from_pin(pin, Nmin=Nmin, Nmax=Nmax)
@@ -78,8 +84,10 @@ def encryptedtext_to_data(encryptedtext: bytearray, password: bytearray, *,
         raise WrongKeyError("WARNING : password and pin can't decrypt the encryptedtext")
     data = decrypt_AES_CBC(ciphertext, derive_key, iv)
     # Deleting from memory all critical data for security
-    delete_bytearray(password)
-    delete_bytearray(pin)
+    if delete_keys:
+        delete_bytearray(password)
+        if pin is not None:
+            delete_bytearray(pin)
     delete_bytearray(derive_key)
     del iterations
     return data
